refactor: purchase 컨트롤러 통합에 따른 문서, 테스트 수정
